# Optimise encodePolyline function