@SiarheiFedartsou 2024-06-16
> Apologies, I may have missed this, but could you highlight where the optimization is?

@DennisOSRM sorry, I left a couple of comments in PR. Let me please know if it is still not clear 🙏🏻